pack: make the host-side file-transfer size cap configurable via SMOLVM_FILE_TRANSFER_MAX_BYTES so a large overlay can be packed without lowering the default DoS bound
