+++ src/agent/mod.rs
@@ -19,7 +19,8 @@ pub use crate::data::network::PortMapping;
 pub use crate::data::resources::VmResources;
 pub use crate::data::storage::HostMount;
 pub use client::{
-    AgentClient, ExecEvent, InteractiveInput, InteractiveOutput, PullOptions, RunConfig,
+    file_transfer_max_total, AgentClient, ExecEvent, InteractiveInput, InteractiveOutput,
+    PullOptions, RunConfig,
 };
 pub use fsnotify_watch::FsNotifyWatcher;
 pub use krun::KrunFunctions;
