+++ src/cli/pack.rs
@@ -1475,7 +1475,7 @@ impl PackCreateCmd {
         dest: &std::path::Path,
         progress_prefix: &str,
     ) -> smolvm::Result<()> {
-        use smolvm_protocol::{AgentRequest, FILE_TRANSFER_MAX_TOTAL};
+        use smolvm_protocol::AgentRequest;
         use std::io::Write;
         use std::time::{Duration, Instant};
 
@@ -1499,6 +1499,7 @@ impl PackCreateCmd {
         let start = Instant::now();
         let mut total_bytes = 0u64;
         let mut last_progress = Instant::now();
+        let cap = smolvm::agent::file_transfer_max_total();
         loop {
             if start.elapsed() > LAYER_EXPORT_TIMEOUT {
                 return Err(Error::agent(
@@ -1519,13 +1520,13 @@ impl PackCreateCmd {
                         // paths do (client.rs): a compromised or buggy guest must not
                         // be able to exhaust the host disk with an endless chunk stream.
                         let next_total = total_bytes.saturating_add(data.len() as u64);
-                        if next_total > FILE_TRANSFER_MAX_TOTAL {
+                        if next_total > cap {
                             let _ = std::fs::remove_file(dest);
                             return Err(Error::agent(
                                 "export layer",
                                 format!(
                                     "guest streamed {} bytes, exceeding the {} byte cap",
-                                    next_total, FILE_TRANSFER_MAX_TOTAL
+                                    next_total, cap
                                 ),
                             ));
                         }
